requirements.txt
discord.py>=2.0
asyncpg
boto3
python-dotenv
psutil